sec2/requirements.txt
alembic==1.13.1
anyio==4.3.0
asgiref==3.5.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
bidict==0.23.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.1.1
django-appconf==1.0.5
django-autocomplete-light==3.11.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==23.5
django-select2==8.0.0
django-selectable==1.4.0
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.96.1
fontawesomefree==6.4.2
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==23.2
pillow==10.2.0
pipdeptree==2.14.0
pipreqs==0.4.13
platformdirs==3.11.0
psutil==5.9.8
psycopg2==2.9.9
pydantic==1.10.14
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
python-engineio==4.9.0
python-multipart==0.0.5
python-socketio==5.11.1
redis==4.6.0
reflex==0.4.1
reflex-hosting-cli==0.1.8
reportlab==4.1.0
requests==2.31.0
rich==13.7.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.27
sqlmodel==0.0.14
sqlparse==0.4.2
starlette==0.27.0
starlette-admin==0.9.0
tabulate==0.9.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2022.2
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.20.0
watchdog==2.3.1
watchfiles==0.19.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
